feat!: enable explicit setting of API base path and prefix
This renders the complicated and error-prone auto-detection obsolete.
It also enables the use of the swagger documentation to directly
interact with the API.

BREAKING CHANGE: URLs are no longer generated by autodetection.
You need to set the environment variable API_HOST_PROTOCOL and,
if the API is at a sub-path, API_BASE_PATH.