feat: allow transactions to be created without an Envelope
This removes the requirement for Incoming and Outgoing transactions to have an Envelope.
With this, “Income” transactions that make money available for budgeting are possible.

At the same time, if a tracking error occurs, reconciling transactions are now possible.
Those either remove money from the budget that does not exist or add money that was not
tracked before (same as an income transaction).

Neither of those two types logically requires an Envelope.