fix: allocation link for envelope month calculation
With this commit, the allocation link for a month without an allocation
points to the collection endpoint. This enables clients to send a POST
request to that endpoint, creating a new allocation, whenever the
budgeted amount for that month is 0.

Before, clients needed to have the URL hardcoded or request it from
the API root, which is not good API design.