fix: importing of overspend handling now accounts for YNAB 4 oddities
In YNAB 4, when the overspendHandling is set to "Confined", it affects all months until
it is explicitly set back to "AffectsBuffer".

EZ on the other hand uses AFFECT_AVAILABLE as default (as does YNAB 4 with "AffectsBuffer")
but only changes to AFFECT_ENVELOPE (= "Confined" on YNAB 4) when explicitly configured for
that month.

This commit introduces parsing logic that accounts for this difference in behaviour when
importing.