fix: add missing overspend mode when importing from YNAB4
When importing from YNAB 4, we account for the different mode handling
of the overspend mode by setting the overspend mode to `AFFECT_ENVELOPE`
for all months up to and including the month when the import happens.

This PR fixes the handling so that it actually includes the `AFFECT_ENVELOPE` setting.