refactor: remove unneeded code
Payees can only be deleted (isTombstone flag) in YNAB 4 when they are
transfer payees (basically a fake payee to be used for transfers to
other accounts).

Since this concept does not exist in Envelope Zero, we do not need to
check for the flag since we skip those payees anyway.