fix!: spent amounts are now negative
With this change, all spent amounts are negative as they should have
been from the beginning. This will make the calculation

  allocation + spend = balance

correct.

BREAKING CHANGE: Sign of spent amounts is flipped to be negative.
If you use the official frontend, upgrade the frontend to 1.9.0 at the same
time as upgrading the backend to 2.0.0.