fix: some boolean fields are NULL
This fixes a bug where boolean fields are NULL for columns that are
created when rows already exist. For these rows, the column is NULL,
which gorm doesn't handle well since we can only query for true of false.

This commit introduces a migration that updates those fields to the default
value, which is always false for booleans.

It's unclear if this is intended or an upstream bug in gorm, therefore I filed
https://github.com/go-gorm/gorm/issues/5968.