docs: finish public API definition
The backend is built with the goal of providing an HTTP API. This means that
the go module itself does not need to be part of the public API to guarantee
stable versioning for users.

If other projects do start relying on it as a go module and deem a more stable
approach necessary, their maintainers are welcome to open an issue to discuss
adding the exported functions and methods to the public API definition for this.

For now however, it will not be part of the public API.