fix: income now recognized in RecentEnvelopes and transaction import
The RecentEnvelopes field for Account resources and the proposed Envelope for
ImportPreviews now also consider income transactions.

Income is represented as the nil UUID in the Envelope ID, meaning that the
RecentEnvelopes field will contain an entry with all zero-values for Income.

This will be improved in API v2 to replace the ID for income with a `null` value.