feat: add accounts API v2 get collection endpoint, improve recentEnvelopes
This adds the `GET /api/v2/accounts` endpoint.

For this endpoint, `recentEnvelopes` now returns only returns the IDs of the
recently used envelopes.

This takes into account the last 50 transactions and returns the 5 most common
envelopes for the account as transaction destination.

Envelopes with the same frequency are returned with the most recently created
envelope first.