feat: add healthz endpoint
This adds the `/healthz` endpoint, which checks the health of the application.
For now, this mainly verifies that the database can be accessed.

Use this endpoint to automatically restart the backend with e.g. a docker compose
health check or Kubernetes probes.

It returns HTTP 204 for success and HTTP 500 with an error message when there is an error.