test: work around codecov token rate limit
With this change, PRs in the repo use the `CODECOV_TOKEN` to upload
coverage to codecov, while PRs from forks will keep using the upload
via the public API, which is suffering from rate limiting issues.

We can't fix the rate limiting issues, but at least we can work around it
for PRs by maintainers for now.