fix: transaction sorting now takes time of day and creation time into account
Before this fix, transactions would be sorted correctly by date, but
transactions later in a day would be after transactions earlier in the day.

When transactions were added for the exact same time (typically when importing),
sorting was undefined. Sorting for transactions at the same time now is "newest first"
by creation time, putting transactions added later earlier in the list.

Since the sorting is "newest first", this was wrong.

Sorting takes place with second precision, meaning that any sub-second differences
are ignored in sorting. This is a limitation of the sqlite `datetime` function.