fix: improve foreign key and not found messages in API v3
This improves the handling of foreign key errors and not found messages in APIv3.

It also moves errors for resource creation to the specific resource, so that
a failure on creation of one resource does not affect all other resources.

With this, the status code for references to non-existing resources is also updated to
404 Not Found, instead of 400 Bad Request.