fix: Month JSON unmarshaling now enforcing year and month only
This fixes an issue where a variable of type types.Month could have
other parts than just year and month set, which was possible when
parsing it from user input.

This never surfaced since this has not been unmarshaled from user input yet
where the input could contain this.

However, this will be the case starting with goals.
Therefore, this bug is fixed now.