refactor!: improve error handling for developers
This massively improves error handling for developers by refactoring
every aspect of the error handling.

This enables handling of errors using the error interface developers are used
to since it is part of the language.

Database error messages are made user friendly directly in the database handling
using gorm callbacks.

This also clarifies that the error messages are not part of the public API.

BREAKING CHANGE: This removes API v3