fix: move to minimal gorm zerolog implementation
This moves to a minimal internal implementation of a logger for gorm that uses zerolog.
This logger uses the ErrResourceNotFound error to skip logs for resources that have not been found.

This change is necessary since we are rewriting the errors in the database callbacks, so the upstream
log handlers don't filter out our "no resource" errors anymore.