feat: add export
This adds functionality to export the whole state of the instance with an API call.
The API call to /v4/export returns a JSON object with some meta information:

* the version used to generate the export
* the creation time

With this change, go 1.22 is required since the tests now use range clauses over integers.

This commit also creates a registry for models so that we do not need to re-implement iteration over all models in the places we need it (database migrations, cleanup, …).