fix: transaction filtering by direction and type
Transaction filtering was incomplete and wrong so far, which is fixed with this change.

There are now two different filters, direction and type.

direction filters by the source/destination accounts being internal or external. It has the following values:

- IN: Transaction from an external to an internal account
- OUT: Transaction from an internal account to an external account
- INTERNAL: Transaction from an internal to an internal account (same as the old direction: TRANSFER)

type filters by the effect a transaction has on the budget. It has the following values

- INCOME: From an off-budget to an on-budget account (same as the old direction: INCOMING)
- SPEND: From an on-budget to an off-budget account (same as the old direction: OUTGOING)
- TRANSFER: From an on-budget to an on-budget account